+++ src/Exception/NotFoundException.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * @copyright 2015-2017 Contentful GmbH
+ * @license   MIT
+ */
+
+namespace Contentful\Exception;
+
+/**
+ * A NotFoundException indicates that the desired resource could not be found on the server.
+ *
+ * @api
+ */
+class NotFoundException extends \RuntimeException
+{
+}